Fixed a few warnings on 64-bit archtectures (namely Alpha). Couldn't fix the  ones in p_file.c though.

+++ trunk/proto2/src/disassem.c
@@ -68,23 +68,23 @@ disassemble(dbref player, dbref program)
                         curr->data.fnumber);
                 break;
             case PROG_ADD:
-                sprintf(buf, "%d: (line %d) ADDRESS: %d", i, curr->line,
+                sprintf(buf, "%d: (line %d) ADDRESS: %zd", i, curr->line,
                         curr->data.addr->data - codestart);
                 break;
             case PROG_TRY:
                 sprintf(buf, "%d: (line %d) TRY: %d", i, curr->line,
-                        curr->data.call - codestart);
+                        (int)(curr->data.call - codestart));
             case PROG_IF:
                 sprintf(buf, "%d: (line %d) IF: %d", i, curr->line,
-                        curr->data.call - codestart);
+                        (int)(curr->data.call - codestart));
                 break;
             case PROG_JMP:
                 sprintf(buf, "%d: (line %d) JMP: %d", i, curr->line,
-                        curr->data.call - codestart);
+                        (int)(curr->data.call - codestart));
                 break;
             case PROG_EXEC:
                 sprintf(buf, "%d: (line %d) EXEC: %d", i, curr->line,
-                        curr->data.call - codestart);
+                        (int)(curr->data.call - codestart));
                 break;
             case PROG_OBJECT:
                 sprintf(buf, "%d: (line %d) OBJECT REF: %d", i, curr->line,
